feat: Add minor imp.

+++ lib/data/models/projedetaymodel/projedetay_model.dart
@@ -9,8 +9,8 @@ part 'projedetay_model.g.dart';
 class ProjeDetay with _$ProjeDetay {
   const factory ProjeDetay({
     @HiveField(0) required bool todoStatus,
-    @HiveField(1) required int kuyuBoy,
-    @HiveField(2) required int kuyuDerinlik,
+    @HiveField(1) required double kuyuBoy,
+    @HiveField(2) required double kuyuDerinlik,
     @HiveField(3) required DateTime createdAt,
   }) = _ProjeDetay;
